Added saveImage.ts
