+++ src/components/Settings/ImageUpload.tsx
@@ -1,6 +1,7 @@
 import config from '@/config'
 import { IMAGE_BOX_SIZE } from '@/data/Constants'
-import { saveImage } from '@/utils/CreateWikiUtils/createWiki'
+import { saveImage } from '@/utils/CreateWikiUtils/saveImage'
+
 import { Box, Spinner, useToast } from '@chakra-ui/react'
 import React, { memo, useRef } from 'react'
 import { Image, NextChakraImageProps } from '../Elements/Image/Image'
